+++ src/deprecated/components/ItemsDialog/ItemsDialog.test.js
@@ -13,8 +13,8 @@ import {
 import {
   itemStatuses,
   requestableItemStatuses,
-} from './constants';
-import { Loading } from './components';
+} from '../../../constants';
+import { Loading } from '../../../components';
 import ItemsDialog, {
   COLUMN_NAMES,
   COLUMN_WIDTHS,
@@ -23,7 +23,7 @@ import ItemsDialog, {
   MAX_HEIGHT,
 } from './ItemsDialog';
 
-jest.mock('./components', () => ({
+jest.mock('../../../components', () => ({
   Loading: jest.fn((props) => (
     <div {...props} />
   )),
@@ -45,6 +45,20 @@ describe('ItemsDialog', () => {
   const testTitle = 'testTitle';
   const testInstanceId = 'testInstanceId';
   const testMutator = {
+    holdings: {
+      GET: jest.fn(() => (new Promise((resolve) => {
+        setTimeout(() => {
+          resolve(
+            [{
+              id: '1',
+            }, {
+              id: '2',
+            }]
+          );
+        });
+      }))),
+      reset: jest.fn(),
+    },
     items: {
       GET: jest.fn(() => (new Promise((resolve) => {
         setTimeout(() => {
@@ -106,6 +120,8 @@ describe('ItemsDialog', () => {
     Paneset.mockClear();
     Loading.mockClear();
     onClose.mockClear();
+    testMutator.holdings.GET.mockClear();
+    testMutator.holdings.reset.mockClear();
     testMutator.items.GET.mockClear();
     testMutator.items.reset.mockClear();
     testMutator.requests.GET.mockClear();
@@ -256,6 +272,7 @@ describe('ItemsDialog', () => {
         });
         const allItemStatuses = Object.values(itemStatuses);
         const newMutator = {
+          ...testMutator,
           items: {
             GET: jest.fn(() => (new Promise((resolve) => {
               setTimeout(() => {