+++ src/deprecated/components/RequestForm/RequestForm.test.js
@@ -17,28 +17,27 @@ import RequestForm, {
   getResourceTypeId,
   ID_TYPE_MAP,
 } from './RequestForm';
-import FulfilmentPreference from './components/FulfilmentPreference';
-import RequesterInformation from './components/RequesterInformation';
-import ItemInformation from './components/ItemInformation';
-import InstanceInformation from './components/InstanceInformation';
+import FulfilmentPreference from '../../../components/FulfilmentPreference';
+import RequesterInformation from '../../../components/RequesterInformation';
 
 import {
   REQUEST_LEVEL_TYPES,
   createModes,
-  RESOURCE_TYPES, REQUEST_LAYERS,
+  REQUEST_LAYERS,
   REQUEST_FORM_FIELD_NAMES,
   DEFAULT_REQUEST_TYPE_VALUE,
   RESOURCE_KEYS,
   REQUEST_OPERATIONS,
-} from './constants';
+} from '../../../constants';
+import { RESOURCE_TYPES } from '../../constants';
 import {
-  getTlrSettings,
   getDefaultRequestPreferences,
   isFormEditing,
   getFulfillmentPreference,
   resetFieldState,
   getRequester,
-} from './utils';
+} from '../../../utils';
+import { getTlrSettings } from '../../utils';
 
 const testIds = {
   tlrCheckbox: 'tlrCheckbox',
@@ -54,8 +53,6 @@ const testIds = {
   closePatronModalButton: 'closePatronModalButton',
   itemDialogCloseButton: 'itemDialogCloseButton',
   itemDialogRow: 'itemDialogRow',
-  itemEnterButton: 'itemEnterButton',
-  instanceEnterButton: 'instanceEnterButton',
 };
 const fieldValue = 'value';
 const idResourceType = 'id';
@@ -65,11 +62,8 @@ const item = {
   barcode: 'itemBarcode',
 };
 
-jest.mock('./utils', () => ({
-  ...jest.requireActual('./utils'),
-  getTlrSettings: jest.fn(() => ({
-    titleLevelRequestsFeatureEnabled: true,
-  })),
+jest.mock('../../../utils', () => ({
+  ...jest.requireActual('../../../utils'),
   getRequestLevelValue: jest.fn(),
   resetFieldState: jest.fn(),
   getDefaultRequestPreferences: jest.fn(),
@@ -77,7 +71,12 @@ jest.mock('./utils', () => ({
   getFulfillmentPreference: jest.fn(),
   getRequester: jest.fn((proxy, selectedUser) => selectedUser),
 }));
-jest.mock('./components/FulfilmentPreference', () => jest.fn(({
+jest.mock('../../utils', () => ({
+  getTlrSettings: jest.fn(() => ({
+    titleLevelRequestsFeatureEnabled: true,
+  })),
+}));
+jest.mock('../../../components/FulfilmentPreference', () => jest.fn(({
   changeDeliveryAddress,
   onChangeAddress,
 }) => {
@@ -100,7 +99,7 @@ jest.mock('./components/FulfilmentPreference', () => jest.fn(({
     </>
   );
 }));
-jest.mock('./components/RequesterInformation', () => jest.fn(({
+jest.mock('../../../components/RequesterInformation', () => jest.fn(({
   findUser,
 }) => {
   const handleChange = () => {
@@ -115,8 +114,8 @@ jest.mock('./components/RequesterInformation', () => jest.fn(({
     />
   );
 }));
-jest.mock('./components/RequestInformation', () => jest.fn(() => <div />));
-jest.mock('./components/ItemInformation', () => jest.fn(({
+jest.mock('../../../components/RequestInformation', () => jest.fn(() => <div />));
+jest.mock('../ItemInformation/ItemInformation', () => jest.fn(({
   findItem,
   triggerValidation,
 }) => {
@@ -133,7 +132,7 @@ jest.mock('./components/ItemInformation', () => jest.fn(({
     />
   );
 }));
-jest.mock('./components/InstanceInformation', () => jest.fn(({
+jest.mock('../InstanceInformation/InstanceInformation', () => jest.fn(({
   findInstance,
   triggerValidation,
 }) => {
@@ -151,7 +150,7 @@ jest.mock('./components/InstanceInformation', () => jest.fn(({
   );
 }));
 jest.mock('@folio/stripes/final-form', () => () => jest.fn((component) => component));
-jest.mock('./PatronBlockModal', () => jest.fn(({
+jest.mock('../../../PatronBlockModal', () => jest.fn(({
   onOverride,
   onClose,
 }) => (
@@ -172,10 +171,10 @@ jest.mock('./PatronBlockModal', () => jest.fn(({
     </button>
   </>
 )));
-jest.mock('./CancelRequestDialog', () => jest.fn(() => <div />));
-jest.mock('./components/TitleInformation', () => jest.fn(() => <div />));
-jest.mock('./ItemDetail', () => jest.fn(() => <div />));
-jest.mock('./ItemsDialog', () => jest.fn(({
+jest.mock('../../../CancelRequestDialog', () => jest.fn(() => <div />));
+jest.mock('../../../components/TitleInformation', () => jest.fn(() => <div />));
+jest.mock('../../../ItemDetail', () => jest.fn(() => <div />));
+jest.mock('../ItemsDialog/ItemsDialog', () => jest.fn(({
   onClose,
   onRowClick,
 }) => {
@@ -202,7 +201,7 @@ jest.mock('./ItemsDialog', () => jest.fn(({
     </>
   );
 }));
-jest.mock('./PositionLink', () => jest.fn(() => <div />));
+jest.mock('../../../PositionLink', () => jest.fn(() => <div />));
 
 describe('RequestForm', () => {
   const labelIds = {
@@ -218,6 +217,7 @@ describe('RequestForm', () => {
     onSetSelectedInstance: jest.fn(),
     onSetSelectedItem: jest.fn(),
     onSetSelectedUser: jest.fn(),
+    onSetInstanceId: jest.fn(),
     onSetIsPatronBlocksOverridden: jest.fn(),
     onSetBlocked: jest.fn(),
     onShowErrorModal: jest.fn(),
@@ -314,14 +314,22 @@ describe('RequestForm', () => {
       });
 
       describe('Initial render', () => {
+        const holding = {
+          holdingsRecords: [
+            {
+              instanceId: 'instanceId',
+            }
+          ],
+        };
         const selectedItem = {
-          instanceId: 'instanceId',
           holdingsRecordId: 'holdingsRecordId',
         };
         let findResource;
 
         beforeEach(() => {
-          findResource = jest.fn().mockResolvedValueOnce(null);
+          findResource = jest.fn()
+            .mockResolvedValueOnce(holding)
+            .mockResolvedValueOnce(null);
 
           const props = {
             ...basicProps,
@@ -361,8 +369,8 @@ describe('RequestForm', () => {
           expect(basicProps.onSetSelectedItem).toHaveBeenCalledWith(undefined);
         });
 
-        it('should get instance information', () => {
-          const expectedArgs = [RESOURCE_TYPES.INSTANCE, selectedItem.instanceId];
+        it('should get instance id if it is not provided', () => {
+          const expectedArgs = [RESOURCE_TYPES.HOLDING, selectedItem.holdingsRecordId];
           const tlrCheckbox = screen.getByTestId(testIds.tlrCheckbox);
 
           fireEvent.click(tlrCheckbox);
@@ -1230,12 +1238,12 @@ describe('RequestForm', () => {
             },
           };
           const itemResult = {
+            totalRecords: 1,
             items: [
               {
                 id: 'itemId',
                 barcode: initialItemBarcode,
                 holdingsRecordId: 'holdingsRecordId',
-                instanceId,
               }
             ],
           };
@@ -1257,6 +1265,13 @@ describe('RequestForm', () => {
           const itemRequestsResult = {
             requests: [],
           };
+          const holdingsRecordResult = {
+            holdingsRecords: [
+              {
+                instanceId,
+              }
+            ],
+          };
           let findResource;
 
           beforeEach(() => {
@@ -1266,6 +1281,7 @@ describe('RequestForm', () => {
               .mockResolvedValueOnce(requestTypesResult)
               .mockResolvedValueOnce(loanResult)
               .mockResolvedValueOnce(itemRequestsResult)
+              .mockResolvedValueOnce(holdingsRecordResult)
               .mockResolvedValue({});
 
             const props = {
@@ -1334,6 +1350,19 @@ describe('RequestForm', () => {
             expect(findResource).toHaveBeenCalledWith(...expectedArgs);
           });
 
+          it('should get information about holdings', () => {
+            const expectedArgs = [
+              RESOURCE_TYPES.HOLDING,
+              itemResult.items[0].holdingsRecordId
+            ];
+
+            expect(findResource).toHaveBeenCalledWith(...expectedArgs);
+          });
+
+          it('should set instance id', () => {
+            expect(basicProps.onSetInstanceId).toHaveBeenCalledWith(holdingsRecordResult.holdingsRecords[0].instanceId);
+          });
+
           it('should handle item barcode field change', () => {
             const expectedArgs = [RESOURCE_TYPES.ITEM, fieldValue, RESOURCE_KEYS.id];
             const itemField = screen.getByTestId(testIds.itemField);
@@ -1355,6 +1384,7 @@ describe('RequestForm', () => {
 
         describe('When item is not found', () => {
           const itemResult = {
+            totalRecords: 0,
             items: [],
           };
           let findResource;
@@ -1401,47 +1431,6 @@ describe('RequestForm', () => {
             const expectedArgs = [RESOURCE_TYPES.REQUEST_TYPES, expect.any(Object)];
 
             expect(findResource).not.toHaveBeenCalledWith(...expectedArgs);
-          });
-        });
-
-        describe('When error during item finding', () => {
-          beforeEach(() => {
-            const props = {
-              ...basicProps,
-              request: undefined,
-              findResource: jest.fn()
-                .mockRejectedValue({}),
-            };
-
-            ItemInformation.mockImplementationOnce(({
-              findItem,
-            }) => {
-              const handleClick = () => {
-                findItem(idResourceType, 'id', false);
-              };
-
-              return (
-                <button
-                  type="button"
-                  data-testid={testIds.itemEnterButton}
-                  onClick={handleClick}
-                >
-                  Enter
-                </button>
-              );
-            });
-
-            renderComponent(props);
-          });
-
-          it('should reset item information', async () => {
-            const itemField = screen.getByTestId(testIds.itemEnterButton);
-
-            fireEvent.click(itemField);
-
-            await waitFor(() => {
-              expect(basicProps.onSetSelectedItem).toHaveBeenCalledWith(null);
-            });
           });
         });
       });
@@ -1493,6 +1482,7 @@ describe('RequestForm', () => {
       const initialItemId = 'itemId';
       const updatedItemId = 'updatedItemId';
       const itemResult = {
+        totalRecords: 0,
         items: [],
       };
 
@@ -1577,8 +1567,13 @@ describe('RequestForm', () => {
           },
         };
         const instanceResult = {
-          id: initialInstanceId,
-          hrid: 'hrid',
+          totalRecords: 1,
+          instances: [
+            {
+              id: initialInstanceId,
+              hrid: 'hrid',
+            }
+          ],
         };
         const requestTypesResult = {
           'Page': [
@@ -1631,8 +1626,8 @@ describe('RequestForm', () => {
 
         it('should set instance information', () => {
           const expectedArgs = [
-            [REQUEST_FORM_FIELD_NAMES.INSTANCE_ID, instanceResult.id],
-            [REQUEST_FORM_FIELD_NAMES.INSTANCE_HRID, instanceResult.hrid]
+            [REQUEST_FORM_FIELD_NAMES.INSTANCE_ID, instanceResult.instances[0].id],
+            [REQUEST_FORM_FIELD_NAMES.INSTANCE_HRID, instanceResult.instances[0].hrid]
           ];
 
           expectedArgs.forEach(args => {
@@ -1653,14 +1648,14 @@ describe('RequestForm', () => {
         it('should get information about open instance requests', () => {
           const expectedArgs = [
             'requestsForInstance',
-            instanceResult.id
+            instanceResult.instances[0].id
           ];
 
           expect(findResource).toHaveBeenCalledWith(...expectedArgs);
         });
 
         it('should set selected instance', () => {
-          expect(basicProps.onSetSelectedInstance).toHaveBeenCalledWith(instanceResult);
+          expect(basicProps.onSetSelectedInstance).toHaveBeenCalledWith(instanceResult.instances[0]);
         });
 
         it('should handle instance id field change', () => {
@@ -1683,7 +1678,10 @@ describe('RequestForm', () => {
       });
 
       describe('When instance is not found', () => {
-        const instanceResult = {};
+        const instanceResult = {
+          totalRecords: 0,
+          instances: [],
+        };
         let findResource;
 
         beforeEach(() => {
@@ -1719,51 +1717,6 @@ describe('RequestForm', () => {
           const expectedArgs = [basicProps.form, REQUEST_FORM_FIELD_NAMES.REQUEST_TYPE];
 
           expect(resetFieldState).not.toHaveBeenCalledWith(...expectedArgs);
-        });
-      });
-
-      describe('When error during instance finding', () => {
-        beforeEach(() => {
-          const props = {
-            ...basicProps,
-            request: undefined,
-            values: {
-              ...basicProps.values,
-              createTitleLevelRequest: true,
-            },
-            findResource: jest.fn()
-              .mockRejectedValue({}),
-          };
-
-          InstanceInformation.mockImplementationOnce(({
-            findInstance,
-          }) => {
-            const handleClick = () => {
-              findInstance('id', false);
-            };
-
-            return (
-              <button
-                type="button"
-                data-testid={testIds.instanceEnterButton}
-                onClick={handleClick}
-              >
-                Enter
-              </button>
-            );
-          });
-
-          renderComponent(props);
-        });
-
-        it('should reset instance information', async () => {
-          const itemField = screen.getByTestId(testIds.instanceEnterButton);
-
-          fireEvent.click(itemField);
-
-          await waitFor(() => {
-            expect(basicProps.onSetSelectedInstance).toHaveBeenCalledWith(null);
-          });
         });
       });
     });