+++ src/deprecated/constants.js
@@ -1,0 +1,8 @@
+export const RESOURCE_TYPES = {
+  ITEM: 'item',
+  INSTANCE: 'instance',
+  USER: 'user',
+  HOLDING: 'holding',
+  REQUEST_TYPES: 'requestTypes',
+};
+