+++ src/index.js
@@ -1,4 +1,5 @@
 import React, { useState } from 'react';
+import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 import { Route, Switch } from 'react-router-dom';
 import ReactRouterPropTypes from 'react-router-prop-types';
@@ -26,6 +27,8 @@ import {
   RequestQueueRoute,
   RequestsRoute,
 } from './routes';
+import DeprecatedRequestsRoute from './deprecated/routes/RequestsRoute/RequestsRoute';
+import DeprecatedRequestQueueRoute from './deprecated/routes/RequestQueueRoute/RequestQueueRoute';
 
 const RequestsRouting = (props) => {
   const {
@@ -72,6 +75,9 @@ const RequestsRouting = (props) => {
       { 'shortcut': 'search', 'label': 'Go to Search & Filter pane' },
       { 'shortcut': 'openShortcutModal', 'label': 'View keyboard shortcuts list' },
     ]);
+  const isEnabledEcsRequests = props.stripes?.config?.enableEcsRequests;
+  const FinalRequestRoute = isEnabledEcsRequests ? RequestsRoute : DeprecatedRequestsRoute;
+  const FinalRequestQueueRoute = isEnabledEcsRequests ? RequestQueueRoute : DeprecatedRequestQueueRoute;
 
   return (
     <>
@@ -104,7 +110,7 @@ const RequestsRouting = (props) => {
           <Switch>
             <Route
               path={`${path}/view/:requestId/:id/reorder`}
-              component={RequestQueueRoute}
+              component={FinalRequestQueueRoute}
             />
             <Route
               path={`${path}/notes/new`}
@@ -120,7 +126,7 @@ const RequestsRouting = (props) => {
             />
             <Route
               path={path}
-              render={() => <RequestsRoute {...props} />}
+              render={() => <FinalRequestRoute {...props} />}
             />
           </Switch>
         </HasCommand>
@@ -141,7 +147,12 @@ const RequestsRouting = (props) => {
 
 RequestsRouting.propTypes = {
   match: ReactRouterPropTypes.match,
-  history: ReactRouterPropTypes.history
+  history: ReactRouterPropTypes.history,
+  stripes: PropTypes.shape({
+    config: PropTypes.shape({
+      enableEcsRequests: PropTypes.bool,
+    }),
+  }).isRequired,
 };
 
 export default RequestsRouting;